simulation/requirements.txt
audioread==3.0.0
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.12
botocore==1.31.12
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
decorator==5.1.1
exceptiongroup==1.1.3
filelock==3.12.2
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
mysql-connector-python==8.1.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
packaging==23.1
paramiko==3.2.0
platformdirs==3.10.0
pluggy==1.3.0
pooch==1.7.0
protobuf==4.21.12
pycparser==2.21
pydub==0.25.1
PyNaCl==1.5.0
pytest==7.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
sympy==1.12
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.16.2
pytest-cov
